+++ CHANGELOG.md
@@ -7,6 +7,7 @@ FEATURES:
 * **New Resource:** `sonatypeiq_application_category` [GH-1]
 * **New Resource:** `sonatypeiq_config_crowd` [GH-1]
 * **New Resource:** `sonatypeiq_config_product_license` [GH-1]
+* **New Resource:** `sonatypeiq_config_saml` [GH-1]
 
 NOTES:
 * Tested against Sonatype IQ Server 194, 195 and 196.